Rename, Moved dummy data, buildForm now module
The main file is now FormBuilder.js
The dummy data is now in test.js.
The buildForm method is now a module.
